Complete user and listing functionality. improve ui

+++ app/Http/Controllers/ListingController.php
@@ -51,6 +51,8 @@ class ListingController extends Controller
             $post_data['logo'] = $request->file('logo')->store('logos', 'public');
         }
 
+        $post_data['user_id'] = auth()->id();
+
         Listing::create($post_data);
 
         return redirect('/')->with('message', 'Listing Created Successfully');
@@ -67,6 +69,12 @@ class ListingController extends Controller
     // Update Listing
     // 
     public function update(Request $request, Listing $listing) {
+
+        // Check if logged in user is owner
+        if( $listing->user_id != auth()->id() ) {
+            abort(403, 'Unauthorized Action');
+        }
+
         $post_data = $request->validate([
             'title' => 'required',
             'company' => 'required',
@@ -90,7 +98,20 @@ class ListingController extends Controller
     // Destroy listing
     // 
     public function destroy(Listing $listing) {
+
+        // Check if logged in user is owner
+        if( $listing->user_id != auth()->id() ) {
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+
+    // 
+    // Manage Page
+    // 
+    public function manage() {
+        return view('listings/manage', ['listings' => auth()->user()->listings()->get()]);
     }
 }
